Remove numeric cast (#5901)

+++ ydb/library/benchmarks/queries/tpcds/pg/q59.sql
@@ -3,13 +3,13 @@
 with wss as
  (select d_week_seq,
         ss_store_sk,
-        sum(case when (d_day_name='Sunday') then ss_sales_price else null::numeric end) sun_sales,
-        sum(case when (d_day_name='Monday') then ss_sales_price else null::numeric end) mon_sales,
-        sum(case when (d_day_name='Tuesday') then ss_sales_price else null::numeric end) tue_sales,
-        sum(case when (d_day_name='Wednesday') then ss_sales_price else null::numeric end) wed_sales,
-        sum(case when (d_day_name='Thursday') then ss_sales_price else null::numeric end) thu_sales,
-        sum(case when (d_day_name='Friday') then ss_sales_price else null::numeric end) fri_sales,
-        sum(case when (d_day_name='Saturday') then ss_sales_price else null::numeric end) sat_sales
+        sum(case when (d_day_name='Sunday') then ss_sales_price else null end) sun_sales,
+        sum(case when (d_day_name='Monday') then ss_sales_price else null end) mon_sales,
+        sum(case when (d_day_name='Tuesday') then ss_sales_price else null end) tue_sales,
+        sum(case when (d_day_name='Wednesday') then ss_sales_price else null end) wed_sales,
+        sum(case when (d_day_name='Thursday') then ss_sales_price else null end) thu_sales,
+        sum(case when (d_day_name='Friday') then ss_sales_price else null end) fri_sales,
+        sum(case when (d_day_name='Saturday') then ss_sales_price else null end) sat_sales
  from {{store_sales}},{{date_dim}}
  where d_date_sk = ss_sold_date_sk
  group by d_week_seq,ss_store_sk
