+++ ydb/library/benchmarks/queries/tpcds/pg/q77.sql
@@ -80,8 +80,8 @@ with ss as
  (select 'store channel' as channel
         , ss.s_store_sk as id
         , sales
-        , coalesce(returns, 0::numeric) as returns
-        , (profit - coalesce(profit_loss,0::numeric)) as profit
+        , coalesce(returns, 0) as returns
+        , (profit - coalesce(profit_loss,0)) as profit
  from   ss left join sr
         on  ss.s_store_sk = sr.s_store_sk
  union all
@@ -96,8 +96,8 @@ with ss as
  select 'web channel' as channel
         , ws.wp_web_page_sk as id
         , sales
-        , coalesce(returns, 0::numeric) returns
-        , (profit - coalesce(profit_loss,0::numeric)) as profit
+        , coalesce(returns, 0) returns
+        , (profit - coalesce(profit_loss,0)) as profit
  from   ws left join wr
         on  ws.wp_web_page_sk = wr.wp_web_page_sk
  ) x
